CR changes: add to comrvErrorHook call to comrvGetStatus
