+++ WD-Firmware/comrv/comrv.c
@@ -27,25 +27,14 @@
 */
 #include "common_types.h"
 #include "comrv.h"
-#include "comrv_config.h"
 #include "comrv_api.h"
-
-/* make sure eviction algorithm defined */
-#ifndef D_COMRV_EVICTION_LRU
- #ifndef D_COMRV_EVICTION_LFU
-  #ifndef D_COMRV_EVICTION_MIX_LRU_LFU
-   /* default eviction algorithm */
-   #define D_COMRV_EVICTION_LRU
-  #endif /* D_COMRV_EVICTION_MIX_LRU_LFU */
- #endif /* D_COMRV_EVICTION_LFU */
-#endif /* D_COMRV_EVICTION_LRU */
 
 /**
 * definitions
 */
 #define D_COMRV_END_OF_STACK                          0xDEAD
 #define D_COMRV_MRU_ITEM                              0xFF
-#define D_COMRV_LRU_ITEM                        0xFF
+#define D_COMRV_LRU_ITEM                              0xFF
 #define D_COMRV_MAX_GROUP_NUM                         0xFFFF
 #define D_COMRV_DWORD_IN_BITS                         32
 #define D_COMRV_ENTRY_LOCKED                          1
@@ -57,7 +46,6 @@
 #define D_COMRV_INVOKE_CALLEE_BIT_0                   1
 #define D_COMRV_RET_CALLER_BIT_0                      0
 #define D_COMRV_NUM_BITS_DWORD                        32
-#define D_COMRV_NUM_OF_CACHE_ENTRIES                  D_COMRV_SIZE_OF_OVL_CACHE_IN_MIN_GROUP_SIZE_UNITS
 #if (D_COMRV_NUM_OF_CACHE_ENTRIES % D_COMRV_NUM_BITS_DWORD)
   #define D_COMRV_EVICT_CANDIDATE_MAP_SIZE            ((D_COMRV_NUM_OF_CACHE_ENTRIES/sizeof(u32_t)) + 1)
 #else
@@ -69,39 +57,6 @@
 #define D_COMRV_GROUP_NOT_FOUND                       0xFFFF
 #define D_COMRV_LAST_MULTI_GROUP_ENTRY                0
 
-/* if no profile was set */
-#if D_COMRV_PROFILE==0
-  #define D_COMRV_PROFILE 1
-#endif /* D_COMRV_PROFILE==0 */
-
-#if D_COMRV_PROFILE == 1
-
-  #ifdef D_COMRV_EVICTION_LRU
-    typedef u08_t lru_t;
-    typedef u16_t lruIndexes_t;
-  #elif defined(D_COMRV_EVICTION_LFU)
-  #elif defined(D_COMRV_EVICTION_MIX_LRU_LFU)
-  #endif /* D_COMRV_EVICTION_LRU */
-
-#elif D_COMRV_PROFILE == 2
-
-  #ifdef D_COMRV_EVICTION_LRU
-    typedef u16_t lru_t;
-    typedef u32_t lruIndexes_t;
-  #elif defined(D_COMRV_EVICTION_LFU)
-  #elif defined(D_COMRV_EVICTION_MIX_LRU_LFU)
-  #endif /* D_COMRV_EVICTION_LRU */
-
-#elif D_COMRV_PROFILE == 3
-
-  #ifdef D_COMRV_EVICTION_LRU
-    typedef u32_t lru_t;
-    typedef u64_t lruIndexes_t;
-  #elif defined(D_COMRV_EVICTION_LFU)
-  #elif defined(D_COMRV_EVICTION_MIX_LRU_LFU)
-  #endif /* D_COMRV_EVICTION_LRU */
-
-#endif /* D_COMRV_PROFILE_1 */
 /**
 * macros
 */
@@ -161,118 +116,6 @@
 /**
 * types
 */
-
-/* comrv stack frame */
-typedef struct comrvStackFrame
-{
-  /* holds return address (caller is non overlay function) or return
-     offset (caller is overlay function) */
-  u32_t uiCallerReturnAddress;
-  /* holds callee address (callee is non overlay function) or callee
-     token (callee is an overlay function) */
-  u32_t uiCalleeToken;
-  /* holds the offset in bytes to the previous stack frame */
-  s16_t ssOffsetPrevFrame;
-#ifdef D_COMRV_MULTI_GROUP_SUPPORT
-  /* if the calleetoken is a multi group token, this field holds the
-     actual loaded token entry in the multi group table */
-  u16_t usCalleeMultiGroupTableEntry;
-#endif /* d_comrv_multi_group_support */
-} comrvStackFrame_t;
-
-/* overlay token fields */
-typedef struct comrvTokenFields
-{
-  /* overlay token indication 0: address; 1: overlay token */
-  u32_t overlayAddressToken:1;
-  /* overlay group ID the function resides in */
-  u32_t overlayGroupID:16;
-  /* data/function offset within the overlay group – 4 bytes granularity */
-  u32_t offset:10;
-  /* caller thunk indication */
-  u32_t thunk:1;
-  /* reserved */
-  u32_t reserved:1;
-  /* specify the cache ID this overlay group belongs to */
-  u32_t cacheID:2;
-  /* multi group indication */
-  u32_t multiGroup:1;
-} comrvTokenFields_t;
-
-/* overlay token */
-typedef union comrvOverlayToken
-{
-  u32_t              uiValue;
-  comrvTokenFields_t stFields;
-} comrvOverlayToken_t;
-
-/* cache entry properties */
-typedef struct comrvPropertiesFields
-{
-  /* entry is ucData or locked by the end user - can't be evicted */
-  u08_t ucLocked:1;
-  /* entry is overlay data */
-  u08_t ucData:1;
-  /* loaded group size in D_COMRV_OVL_GROUP_SIZE_MIN granularity */
-  u08_t ucSizeInMinGroupSizeUnits:4;
-  u08_t reserved:2;
-} comrvPropertiesFields_t;
-
-/* cache entry */
-typedef union comrvEntryProperties
-{
-  u08_t                   ucValue;
-  comrvPropertiesFields_t stFields;
-} comrvEntryProperties_t;
-
-#ifdef D_COMRV_EVICTION_LRU
-/* overlay cache entry */
-typedef union comrvEvictionLru
-{
-   struct comrvEvictionLruFields_t
-   {
-      /* bidirectional linked list - index of previous LRU item */
-      lru_t     typPrevLruIndex;
-      /* bidirectional linked list - index of next LRU item */
-      lru_t     typNextLruIndex;
-   } stFields;
-   /* value of both lru fields in a single value */
-   lruIndexes_t typValue;
-} comrvEvictionLru_u;
-#elif defined(D_COMRV_EVICTION_LFU)
-#elif defined(D_COMRV_EVICTION_MIX_LRU_LFU)
-#endif /* D_COMRV_EVICTION_LRU */
-
-/* overlay cache entry */
-typedef struct comrvCacheEntrycomrvCacheEntry
-{
-#ifdef D_COMRV_EVICTION_LRU
-   comrvEvictionLru_u    unLru;
-#elif defined(D_COMRV_EVICTION_LFU)
-#elif defined(D_COMRV_EVICTION_MIX_LRU_LFU)
-#endif /* D_COMRV_EVICTION_LRU */
-  /* entry properties */
-  comrvEntryProperties_t unProperties;
-  /* loaded token */
-  comrvOverlayToken_t    unToken;
-  /* fixed RAM address that this cache entry refers to */
-  void*                  pFixedEntryAddress;
-} comrvCacheEntry_t;
-
-/* comrv control block */
-typedef struct comrvCB
-{
-#ifdef D_COMRV_EVICTION_LRU
-  /* holds the cache entry index of the LRU item */
-  lru_t             ucLruIndex;
-  /* holds the cache entry index of the MRU item */
-  lru_t             ucMruIndex;
-#elif defined(D_COMRV_EVICTION_LFU)
-#elif defined(D_COMRV_EVICTION_MIX_LRU_LFU)
-#endif /* D_COMRV_EVICTION_LRU */
-  /* the cache entries */
-  comrvCacheEntry_t stOverlayCache[D_COMRV_NUM_OF_CACHE_ENTRIES];
-} comrvCB_t;
 
 /**
 * local prototypes
@@ -853,13 +696,14 @@ static void comrvUpdateCacheEvectionParams(u08_t ucEntryIndex)
 /**
 * get comrv status
 *
-* @param none
+* @param pComrvStatus - function output
 *
 * @return none
 */
-void comrvGetStatus(void)
+void comrvGetStatus(comrvStatus_t* pComrvStatus)
 {
-   // TODO: check comrv stack
+   pComrvStatus->pComrvCB    = &g_stComrvCB;
+   pComrvStatus->pComrvStack = g_stComrvStackPool;
 }
 
 /**
